1학기 cloud9 총정리/django_friday/workspace/Django_API_Vue/requirements.txt
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==2.1.8
django-allauth==0.39.1
django-appconf==1.0.3
django-bootstrap4==0.0.8
django-crispy-forms==1.7.2
django-imagekit==4.0.2
django-rest-swagger==2.2.0
django-test-plus==1.1.1
djangorestframework==3.9.2
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
oauthlib==3.0.1
openapi-codec==1.3.2
pilkit==2.0
Pillow==6.0.0
python3-openid==3.1.0
pytz==2019.1
requests==2.21.0
requests-oauthlib==1.2.0
simplejson==3.16.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.2
